Refactors translation suggestions
Uses native language toasts for 8 most popular languages on US
